mcga: add argument in prctl to promote or collapse pages
Currently, the supported prctl call to enable mcga is
    prctl(47, 1, 0, 0, 0);

this call will always promote or collapse by copying
depending on a static policy decided during compilation.
We need to re-compilte the kernel every time to measure
the performance of various benchmarks when khugepaged is
configured for promotion/collapse.

To configure this runtime, the prctl call now accepts extra
parameter.
    prctl(47, 1, 0, 0, 0); - will collapse by copy
    prctl(47, 1, 1, 0, 0); - wiill promote